feat(host/parents-gateway): consume PG's split Posts/Groups remotes
Parents Gateway's team split their single `pg/App` export into two
separate federation modules, `pg/Posts` and `pg/Groups`, exposed at
their own top-level routes rather than one shared entry point. Update
the host to match: extract the remote-loading/error-boundary wiring
into a shared helper, load `pg/Posts` at /posts (replacing the
no-longer-existing pg/App) and pg/Groups at a new /groups route.
