+++ apps/host/src/containers/GroupsGatewayView.tsx
@@ -1,0 +1,3 @@
+import { createParentsGatewayRemoteView } from '~/containers/parents-gateway-remote';
+
+export const GroupsGatewayView = createParentsGatewayRemoteView('pg/Groups', 'Groups');
